fix(frontend): レビュー指摘の修正（カスタムサウンド管理のルート分離・エラー時のフォールバック等）
- feat/195-instance-custom-sounds のレビュー指摘 4 件に対応（CHANGELOG を除く）
- en-US.yml に fileAlreadyUsed / fileAlreadyUsedTitle を追加（空ダイアログ防止）
- カスタムサウンド管理UIを /settings/custom-sounds へ移設し、canManageCustomSounds
  ロール保持者（非 admin/mod）でも管理可能に（iAmModerator ガードの外へ）
- tl-note-notification.ts の get-custom-sounds 取得失敗時に空配列へフォールバック
- sounds.sound.vue で url:null の破損サウンドを選択肢から除外
